+++ ft_printf.h
@@ -25,7 +25,7 @@ int		print_c(int c);
 int		print_str(char *str);
 int		print_num(int n);
 int		print_unum(unsigned int n);
-int		print_hex(long int num, char arg);
+int		print_hex(unsigned int num, char arg);
 int		print_point(void *pointer);
 int		fill_ptr(char **pointer, uintptr_t num);
 int		prefix(char **pointer);
